Rename to break
This is it, folks! We have a great name with a ring to it here. 💍

```ruby
binding.break # It is!
```

To be honest, I never liked the boogah name, but sticked to it for a
while. I wanna push this project further, though, and that requires a
better name to market it. Break checks all the boxes. ✅

✓ It is related to a breakpoint, a commonly used debugger term.
✓ It's much easier to remember, search and type than `boogah`,
✓ It sounds better overall.