Overhaul both IRB and Pry sessions integrations
This change overhauls the session management system and solves two big
pending problems:

1. Calling `next` in IRB is no longer a `SyntaxError` but a `break` next
   command call.

2. Stepping into `binding.irb` or `binding.pry` does not start a new
   session but jumps over the call. Stepping into a call is quite
   problematic the stepping skips internal frames for `break` and ends
   up in the frames the people are not expecting. Let's not allow it in
   the first place. Fixing it is rather finicky at the moment, but it
   will get better with time.

Other than the major changes, the IRB integration is now separated like
the Pry one. This implies that IRB is not special rather it's just the
default frontend.

The other change is that `binding.break` is gone now. We don't need it.
Maybe with a remote frontend, but who knows. Let's wait it out!