Change the debugging context outside of TracePoint handlers
If the frontend is started inside of a TracePoint handler, no TracePoint
events will be issued in that context. This is a problem with Zeitwerk
autoloading, but it's also a general one. TracePoint is used for more
than debugging tools now and it's worth it to investigate how we can
start debugging sessions outside of it's handlers as we cannot expect
good upstream (think Ruby VM) support for this kind of thing.