Drop the session tracking code
It's not working at the moment and I need a better way to trace the
session execution that is not so hacky and global state dependent.

Break execution is imitated by the Ruby VM calling-out the TracePoint
handlers when an event is hit. To make it even more complicated, I'm
jumping off to a Fiber when a session context is switched, so I can run
TracePoint code in a debugging session, and globals are not a good way
to indicate when tracing is happening and when not...